+++ python_packages/jupyter_lsp/jupyter_lsp/specs/__init__.py
@@ -22,6 +22,7 @@ from .vscode_css_languageserver import VSCodeCSSLanguageServer
 from .vscode_html_languageserver import VSCodeHTMLLanguageServer
 from .vscode_json_languageserver import VSCodeJSONLanguageServer
 from .yaml_language_server import YAMLLanguageServer
+from .zuban import ZubanLanguageServer
 
 basedpyright = BasedPyrightLanguageServer()
 bash = BashLanguageServer()
@@ -42,3 +43,4 @@ ts_old = JavascriptTypescriptLanguageServer()
 ts = TypescriptLanguageServer()
 sql = SQLLanguageServer()
 yaml = YAMLLanguageServer()
+zuban = ZubanLanguageServer()